Fix -Wsign-compare

git-svn-id: http://opkg.googlecode.com/svn/trunk@702 e8e0d7a0-c8d9-11dd-a880-a1081c7ac358

+++ libopkg/opkg_message.c
@@ -78,7 +78,7 @@ opkg_message (message_level_t level, const char *fmt, ...)
 {
 	va_list ap;
 
-	if (opkg_config->verbosity < level)
+	if (opkg_config->verbosity < (int)level)
 		return;
 
 	if (opkg_config->opkg_vmessage) {
